gpu mem fixes, json parsing, req file update

+++ src/script/scripter.py
@@ -40,7 +40,7 @@ class LLMFacade:
             self.tokenizer = AutoTokenizer.from_pretrained(self.model_id)
             self.model = AutoModelForCausalLM.from_pretrained(
                 self.model_id,
-                torch_dtype=self.dtype,
+                dtype=self.dtype,
                 device_map=self.device_map
             )
